Update patch set 1
Patch Set 1:

While this method works on the surface, it is not the best way to do it. (BTW you need to check if FM radio is enabled or not because setFmVolume is always called regardless of if the radio is enabled or not).

A problem with this method is that it does not work with Voodoo Sound's extended volume control. The stock FM does work with that because it uses the WM8994 volume gain instead of the FM chipset's. I'm working on a kernel patch which should fix both the volume gain and the audio routing problem. (The audio routing problem comes up when you use the music player, turn it off, turn on the FM radio and the sound doesn't come until you toggle speaker/headset.)

Patch-set: 1