Update patch set 2
Patch Set 2:

You may be right, however, IMHO we do need to support the camer shutter sound toggle. It would be annoying to have this everywhere but on samsung devices.

As discussed on IRC yesterday, I would be fine with using the OSS libcameraservice - yet I'm unsure about why it was used in the first place. I would be happy to use the OSS one, but I'm using the camera way too seldom to spot bugs or missing features.

PS: That was least hackish solution I could think of. I had several other ideas... ;-)

Patch-set: 2