Update patch set 2
Patch Set 2:

OSS libcameraservice.so didn't work on captivate for some reason. Getting it to work there would be the non-hackish solution for this, but I have no captivate. The other acceptable solution would be to hex (using assembly) the proprietary library to respect the two properties. But that's probably harder than the first solution.

Patch-set: 2