Update patch set 2
Patch Set 2:

Patching the libcameraservice.so to check the properties' value there would be possible and even less hackish, right. Problem is that there's no symbol exported nearby playSound() where I could inject code. Patching it manually would require redistributing the patched .so along with the code, which seems unacceptable to me.

So if this is a captivate issue only it would be fine to use the current OSS camerservice on other aries devices?

Patch-set: 2