Update patch set 2
Patch Set 2:

Yeah, but that would just be being lazy! To have it activate it then shut it down if it's not supposed to be, would be bad.

I have seen that commit before, and that commit actually doesn't restrict it to incall only, it just replaces the prevMode check with a mode check for either ringtone or incall.

"Crespo actually does setCallClockSync on MODE_RINGTONE, which didn't quite work out for us as reported by some users."
You knew it didn't work for us, yet we still do it? Our libaudio code-base is almost identical to crespo, save FM radio components.

Frankly, the Vibrant can live without that function call (builds I uploaded without the call worked fine) but I'd rather keep it. (our broken E911 libaudio for Vibrant didn't have it)

Patch-set: 2