Update patch set 2
Patch Set 2:

No, I meant this.

        if (prevMode != AudioSystem::MODE_IN_CALL && mMode == AudioSystem::MODE_IN_CALL)
        {
            if ((!mActivatedCP) && (mSecRilLibHandle) && (connectRILDIfRequired() == OK)) {

mActivatedCP keeps track of whether the CP clock is activated or not. If it's already activated (like MODE_IN_CALL -> MODE_IN_CALL, which really shouldn't happen), then that line is skipped anyway so the prevMode part isn't necessary.

As for our libaudio being similar to crespo, it's because it's a copy of the crespo's libaudio (by Google/Samsung), which happens to work. I'm reluctant to make changes I don't understand (such as adding a stop call clock) due to the fear of breaking more things.

"You knew it didn't work for us, yet we still do it?"

The change in question didn't work out, the revert worked for the user reporting it. I didn't investigate further.

Patch-set: 2