Update patch set 1
Patch Set 1: Verified; Looks good to me, approved

Warren, if the epic uses yas529, it's probably affected by this bug as well. You can test by reading /sys/class/input/event2/device/delay (or whatever the correct path is on your device) and see if it changes as it should when setting different delays with an app like "Sensor List" or not.

Patch-set: 1