Update patch set 1
Patch Set 1:

Fascinate users are extremely picky apparently...

The defaults with the corresponding kernel patch are fine for me personally, but there seems to be a widely different opinion from users.

As this is an extremely popular feature in the glitch kernel, I thought it would be useful in CM, since the original author signed off on it without reservation.

Patch-set: 1