Update patch set 1
Patch Set 1:

There are a few reasons I took this approach instead.

Firstly, many CM users flash the Glitch kernel specifically for this feature. This allows them to remove a step, and keep the same features. This feature has been around for almost a year, so users that take advantage of it already know what their preferred settings are. Adding the option to AriesParts removes the dependency on the init.d scripts that are currently bundled with the Glitch kernel.

Secondly, there are many bugs present in the current in-call audio code for the fascinate, that have been there for some time. Efpophis has put much work into this feature, and I don't wish to duplicate the work. Due to the fact that microphone gain cannot be manipulated in android, and the correlation between it and in-call volume, I've burned much time in the past trying (and failing) to come up with a "standardized" solution.

Thirdly, the only device I have physical access to now is my wife's fascinate. That gives me limited development time/opportunity ;) Due to the fact I've not had much luck in the past coming up with an independent solution when I still had a personal device, it will take much more "borrowing" of her phone to come up with one now.

Patch-set: 1