Update patch set 3
Patch Set 3:

No, the devices weren't "running fine on their original configuration".

Ever since CM7, /data/data was symlinked to /datadata on the OneNAND for performance reason. That came with its own problem being that /datadata is small that user will see out of space warning for that partition, but the user will see that /data has plenty of partition. Samsung also did something similar with their stock ROM moving databases onto /dbdata.

In Jellybean, the /system partition was larger and /datadata was removed to make space for it. We were running without it for a while but that caused a lot of unbearable lag for a number of users (me included). 

If you use systrace, you can see that writing a block or two can easily take half a second.

Patch-set: 3