Update patch set 3
Patch Set 3:

Got a suggestion for a good workload to test with systrace?

I'm fairly certain the Infuse isn't affected but it would be beneficial to confirm, as later this week I'm going to resume Infuse work and I think I may need to make sure it avoids using anything from aries-common that was affected by this change.  It would be interesting to determine if the performance issue is caused the eMMC chip in use, or if it's some problem with the aries kernel MMC driver.

As I mentioned previously, there's at least one Hummingbird device that is pure eMMC with no OneNAND, and so it is at least possible with some chip/driver combo to achieve good performance.

Patch-set: 3