Update patch set 4
Patch Set 4:

Sorry, uploaded the wrong log. This is the correct one:
http://pastebin.com/jPDXEm0n

The log shows that I change ISO from auto to 1600 to 800, then take a pic with 800. After resetting to default values, I press the settings button again, and then there is a range of errors followed by FC of the Camera app.

Patch-set: 4