Update patch set 5
Patch Set 5:

These are the layout values I'm playing with: https://gist.github.com/4188548

Sacrificing unlock height so that widget can be displayed properly, reducing left margin so that the "pqrs" and "wxyz"  on pin lock can be displayed properly. Pin unlock screen still broken because Google decided to display an empty widget on top, leaving very little space to display both the sim icon and the keypad.

Integrating these into the framework will also be a bit of a challenge because Android provides override for larger screen, not smaller screens, so we'll have to change the default and override it so that larger screen display it properly.

We simply don't have enough space and Google didn't have hdpi devices in mind at all :(

Patch-set: 5