Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Get rid of the non-free playlpm binary and use charger instead with free bitmaps (svg source is provided as well).

I also updated the recovery graphics (because that's what charger uses) because it was RGB565 and the images looked bad (the gradients were not very pretty).

Note that this must come with changes to the kernel, since the init script was doing some ugly workaround moving lpm.rc to init.rc. Let the real init binary handle that cleanly!

Patch-set: 1