Update patch set 1
Patch Set 1:

Did you check if this works when called from within init (make sure to erase the old data file before testing)? The last time I did this, bdaddr_read was called before the property was set. We might need to change when this is called, and prop reading works on GSM as well, so no need for ifdef.

Patch-set: 1