Update patch set 1
Patch Set 1:

The version I was using previously wrote the information to /data/.bdaddr, so I made sure to change that in my device tree. I tested it, and it works fine.

In init.aries.rc, the service isn't started until the system property is available, which fortunately is early enough in the boot process for it to work correclty.

Patch-set: 1