Update patch set 1
Patch Set 1:

I took a look. I'm afraid it'd change the behavior from "backlight on when button is touched" to "backlight on when screen is touched", and many users will complain about it (and I'm not joking... I have another device with few users and many are complaining about AOSP vs Samsung backlight behavior).

Not to mention that the backlight automatically gets turned on when the button is touched anyway (at least on aries, by the button's firmware), so the "disable" option won't work properly.

And finally, I'm lazy.

Patch-set: 1